requirements.txt
langchain==0.0.323
langchain-community==0.0.12
langchain-core==0.1.10
langchain-experimental==0.0.34
langdetect==1.0.9
langsmith==0.0.63
sounddevice~=0.4.6
pyttsx3~=2.90
playsound~=1.2.2
scipy~=1.8.0
gTTS~=2.5.0
streamlit~=1.28.1
huggingface_hub==0.16.4
SpeechRecognition==3.10.1
unstructured==0.10.16
protobuf==4.21.12
sentence-transformers==2.2.2
chromadb==0.3.29
duckduckgo-search==4.2
wikipedia==1.4.0
numexpr==2.8.8
unstructured-inference==0.6.5
unstructured.pytesseract==0.3.12
PyAudio
streamlit-audiorecorder
pycairo
PyGObject
aiortc
pygame